WGiEdit: Delete song option now deletes empty folders as well. WGiEdit: Deleting song audio is now less problematic, as the audio is released from FMOD's control first. Game Log is now saved to the WGiBeat folder. It can be disabled if desired from the options menu. WARNING: Due to FMOD changes, the crossfader is highly unstable and causes crashes. Will be fixed in a future commit.

+++ WindowsGame1/WindowsGame1/AudioSystem/SongManager.cs
@@ -252,10 +252,17 @@ namespace WGiBeat.AudioSystem
                     RemoveSong(existingSongFile);
                 }
                 File.Delete(song.Path + "\\" + song.DefinitionFile);
-                if (deleteAudio)
+                if ((deleteAudio) && File.Exists(song.Path + "\\" + song.AudioFile))
                 {
+                    AudioManager.ReleaseSound(song.Path + "\\" + song.AudioFile);
                     File.Delete(song.Path + "\\" + song.AudioFile);
+
+                    if (Directory.GetFiles(song.Path).Length == 0)
+                    {
+                        Directory.Delete(song.Path);
+                    }
                 }
+               
                 return "";
             }
             catch (Exception ex)
